refactor(acp): sever nori-acp's dependency on codex-core
Move the harness-domain helpers nori-acp consumed from core into nori-acp
itself: user_notification, custom_prompts (dead default_prompts_dir
dropped), parse_command + shell/bash/powershell, compact constants (+
templates), and create_patch_with_context (new patch.rs, out of util.rs).
Hoist McpServerConfig/McpServerTransportConfig into
codex_protocol::config_types; core keeps an intra-crate re-export shim.
nori-acp's Cargo.toml no longer lists codex-core; cargo tree confirms the
edge is gone. tui/cli reach the moved helpers via nori_acp::.

Slice E of docs/specs/crate-layering.md §6.
