+++ nori-rs/acp/src/backend/session.rs
@@ -301,7 +301,7 @@ impl AcpBackend {
         let (prompt_result_tx, prompt_result_rx) = mpsc::channel(128);
         let use_native_notifications =
             config.os_notifications == crate::config::OsNotifications::Enabled;
-        let user_notifier = Arc::new(codex_core::UserNotifier::new(
+        let user_notifier = Arc::new(crate::UserNotifier::new(
             config.notify.clone(),
             use_native_notifications,
         ));
