+++ nori-rs/acp/templates/compact/prompt.md
