+++ nori-rs/core/tests/common/Cargo.toml
@@ -11,6 +11,7 @@ path = "lib.rs"
 anyhow = { workspace = true }
 assert_cmd = { workspace = true }
 codex-core = { workspace = true }
+nori-acp = { workspace = true }
 notify = { workspace = true }
 regex-lite = { workspace = true }
 shlex = { workspace = true }
