+++ nori-rs/core/tests/common/lib.rs
@@ -59,7 +59,7 @@ pub fn sandbox_network_env_var() -> &'static str {
 }
 
 pub fn format_with_current_shell(command: &str) -> Vec<String> {
-    codex_core::shell::default_user_shell().derive_exec_args(command, true)
+    nori_acp::shell::default_user_shell().derive_exec_args(command, true)
 }
 
 pub fn format_with_current_shell_display(command: &str) -> String {
