+++ nori-rs/tui/src/exec_command.rs
@@ -1,8 +1,8 @@
 use std::path::Path;
 use std::path::PathBuf;
 
-use codex_core::parse_command::extract_shell_command;
 use dirs::home_dir;
+use nori_acp::parse_command::extract_shell_command;
 use shlex::try_join;
 
 pub(crate) fn escape_command(command: &[String]) -> String {
